+++ pywatts/core/step.py
@@ -1,8 +1,12 @@
 import logging
 import time
+import warnings
+
+import pandas as pd
 from typing import Optional, Dict, Union, Callable, List
 
 import cloudpickle
+import pandas as pd
 import xarray as xr
 
 from pywatts.callbacks import BaseCallback
@@ -13,6 +17,7 @@ from pywatts.core.computation_mode import ComputationMode
 from pywatts.core.exceptions.not_fitted_exception import NotFittedException
 from pywatts.core.filemanager import FileManager
 from pywatts.core.result_step import ResultStep
+from pywatts.utils._xarray_time_series_utils import _get_time_indexes
 
 logger = logging.getLogger(__name__)
 
@@ -36,8 +41,14 @@ class Step(BaseStep):
     :type callbacks: List[Union[BaseCallback, Callable[[Dict[str, xr.DataArray]], None]]]
     :param condition: A callable which checks if the step should be executed with the current data.
     :type condition: Callable[xr.DataArray, xr.DataArray, bool]
-    :param train_if: A callable which checks if the train_if step should be executed or not.
-    :type train_if: Callable[xr.DataArray, xr.DataArray, bool]
+    :param refit_condition: A ConditionObject which checks if the step should be refitted or not.
+    :type refit_condition: ConditionObject
+    :param lag: Needed for online learning. Determines what data can be used for retraining.
+            E.g., when 24 hour forecasts are performed, a lag of 24 hours is needed, else the retraining would
+            use future values as target values.
+    :type lag: pd.Timedelta
+    :param retrain_batch: Needed for online learning. Determines how much data should be used for retraining.
+    :type retrain_batch: pd.Timedelta
     """
 
     def __init__(self, module: Base, input_steps: Dict[str, BaseStep], file_manager, *,
@@ -46,16 +57,25 @@ class Step(BaseStep):
                  callbacks: List[Union[BaseCallback, Callable[[Dict[str, xr.DataArray]], None]]] = [],
                  condition=None,
                  batch_size: Optional[None] = None,
-                 train_if=None):
-
+                 refit_condition=None,
+                 retrain_batch=pd.Timedelta(hours=24),
+                 lag=pd.Timedelta(hours=24)):
         super().__init__(input_steps, targets, condition=condition,
-                         computation_mode=computation_mode, name=module.name)
+                         computation_mode=computation_mode, name=module.name, lag=lag)
         self.file_manager = file_manager
         self.module = module
+        self.retrain_batch = retrain_batch
         self.callbacks = callbacks
         self.batch_size = batch_size
-        self.train_if = train_if
+        if self.current_run_setting.computation_mode not in [ComputationMode.Refit] and refit_condition is not None:
+            message = "You added a refit_condition without setting the computation_mode to refit." \
+                      " The conditoin will be ignored."
+            warnings.warn(message)
+            logger.warn(message)
+        self.lag = lag
+        self.refit_condition = refit_condition
         self.result_steps: Dict[str, ResultStep] = {}
+        self.retrain_batch = retrain_batch
 
     def _fit(self, inputs: Dict[str, BaseStep], target_step):
         # Fit the encapsulate module, if the input and the target is not stopped.
@@ -64,13 +84,15 @@ class Step(BaseStep):
     def _callbacks(self):
         # plots and writs the data if the step is finished.
         for callback in self.callbacks:
+            dim = _get_time_indexes(self.buffer)[0]
+
+            to_plot = self.buffer
+            if self.current_run_setting.online_start is not None:
+                to_plot = {k: self.buffer[k][self.buffer[k][dim] >= self.current_run_setting.online_start] for k in
+                           self.buffer.keys()}
             if isinstance(callback, BaseCallback):
                 callback.set_filemanager(self.file_manager)
-            if isinstance(self.buffer, xr.DataArray) or isinstance(self.buffer, xr.Dataset):
-                # DEPRECATED: direct DataArray or Dataset passing is depricated
-                callback({"deprecated": self.buffer})
-            else:
-                callback(self.buffer)
+            callback(to_plot)
 
     def _transform(self, input_step):
         if isinstance(self.module, BaseEstimator) and not self.module.is_fitted:
@@ -96,11 +118,11 @@ class Step(BaseStep):
                 condition = cloudpickle.load(pickle_file)
         else:
             condition = None
-        if stored_step["train_if"]:
-            with open(stored_step["train_if"], 'rb') as pickle_file:
-                train_if = cloudpickle.load(pickle_file)
+        if stored_step["refit_condition"]:
+            with open(stored_step["refit_condition"], 'rb') as pickle_file:
+                refit_condition = cloudpickle.load(pickle_file)
         else:
-            train_if = None
+            refit_condition = None
         callbacks = []
         for callback_path in stored_step["callbacks"]:
             with open(callback_path, 'rb') as pickle_file:
@@ -109,7 +131,7 @@ class Step(BaseStep):
             callbacks.append(callback)
 
         step = cls(module, inputs, targets=targets, file_manager=file_manager, condition=condition,
-                   train_if=train_if, callbacks=callbacks, batch_size=stored_step["batch_size"])
+                   refit_condition=refit_condition, callbacks=callbacks, batch_size=stored_step["batch_size"])
         step.default_run_setting = RunSetting.load(stored_step["default_run_setting"])
         step.current_run_setting = step.default_run_setting.clone()
         step.id = stored_step["id"]
@@ -118,52 +140,75 @@ class Step(BaseStep):
 
         return step
 
+    def refit(self, start: pd.Timestamp, end: pd.Timestamp):
+        if self.current_run_setting.computation_mode in [ComputationMode.Refit] and isinstance(self.module,
+                                                                                               BaseEstimator):
+            if self.refit_condition:
+                condition_input = {key: value.step.get_result(start, end) for key, value in
+                                   self.refit_condition.kwargs.items()}
+                if self.refit_condition.evaluate(**condition_input):
+                    # TODO should the same data be used for refitting and for calling the refit_condition condition?
+                    #      NOPE -> Make it more flexible... Perhaps something for issue 147
+                    refit_input = self._get_input(end - self.retrain_batch, end)
+                    refit_target = self._get_target(end - self.retrain_batch, end)
+                    start_time = time.time()
+                    summary = self.module.refit(**refit_input, **refit_target)
+                    self.refit_time.set_kv(f"Refit at position {end}", time.time() - start_time)
+
     def _compute(self, start, end):
         input_data = self._get_input(start, end)
         target = self._get_target(start, end)
         if self.current_run_setting.computation_mode in [ComputationMode.Default, ComputationMode.FitTransform,
-                                                         ComputationMode.Train] and (
-                not self.train_if or self.train_if(input_data, target)):
+                                                         ComputationMode.Train]:
             # Fetch input_data and target data
-            start_time = time.time()
             if self.batch_size:
                 input_batch = self._get_input(end - self.batch_size, end)
                 target_batch = self._get_target(end - self.batch_size, end)
+                start_time = time.time()
                 self._fit(input_batch, target_batch)
+                self.training_time.set_kv("", time.time() - start_time)
             else:
+                start_time = time.time()
                 self._fit(input_data, target)
-            self.training_time.set_kv("", time.time() - start_time)
+                self.training_time.set_kv("", time.time() - start_time)
         elif self.module is BaseEstimator:
             logger.info("%s not fitted in Step %s", self.module.name, self.name)
 
         start_time = time.time()
         result = self._transform(input_data)
+
         self.transform_time.set_kv("", time.time() - start_time)
-        return result
+        result_dict = {}
+        for key, res in result.items():
+            index = res.indexes[_get_time_indexes(result)[0]]
+            start = max(index[0], start.to_numpy())
+            result_dict[key] =  res.sel(**{_get_time_indexes(res)[0]: index[(index >= start)]})
+        return result_dict
 
     def _get_target(self, start, batch):
         return {
-            key: target.get_result(start, batch) for key, target in self.targets.items()
+            key: target.get_result(start - self.module.get_min_data(), batch) for key, target in self.targets.items()
         }
 
     def _get_input(self, start, batch):
         return {
-            key: input_step.get_result(start, batch) for key, input_step in self.input_steps.items()
+            key: input_step.get_result(start - self.module.get_min_data(), batch) for key, input_step in
+            self.input_steps.items()
         }
 
     def get_json(self, fm: FileManager):
         json = super().get_json(fm)
         condition_path = None
-        train_if_path = None
+        refit_condition_path = None
         callbacks_paths = []
         if self.condition:
             condition_path = fm.get_path(f"{self.name}_condition.pickle")
             with open(condition_path, 'wb') as outfile:
                 cloudpickle.dump(self.condition, outfile)
-        if self.train_if:
-            train_if_path = fm.get_path(f"{self.name}_train_if.pickle")
-            with open(train_if_path, 'wb') as outfile:
-                cloudpickle.dump(self.train_if, outfile)
+        if self.refit_condition:
+            refit_condition_path = fm.get_path(f"{self.name}_refit_condition.pickle")
+            with open(refit_condition_path, 'wb') as outfile:
+                cloudpickle.dump(self.refit_condition, outfile)
         for callback in self.callbacks:
             callback_path = fm.get_path(f"{self.name}_callback.pickle")
             with open(callback_path, 'wb') as outfile:
@@ -171,9 +216,25 @@ class Step(BaseStep):
             callbacks_paths.append(callback_path)
         json.update({"callbacks": callbacks_paths,
                      "condition": condition_path,
-                     "train_if": train_if_path,
+                     "refit_condition": refit_condition_path,
                      "batch_size": self.batch_size})
         return json
+
+    def refit(self, start: pd.Timestamp, end: pd.Timestamp):
+        """
+        Refits the module of the step.
+        :param start: The date of the first data used for retraining.
+        :param end: The date of the last data used for retraining.
+        """
+        if self.current_run_setting.computation_mode in [ComputationMode.Refit] and isinstance(self.module,
+                                                                                               BaseEstimator):
+            if self.refit_condition:
+                input_data = self._get_input(start, end)
+                target = self._get_target(start, end)
+                if self.refit_condition(input_data, target):
+                    refit_input = self._get_input(end - self.retrain_batch, end)
+                    refit_target = self._get_target(end - self.retrain_batch, end)
+                    self.module.refit(**refit_input, **refit_target)
 
     def get_result_step(self, item: str):
         if item not in self.result_steps: