+++ CHANGELOG.md
@@ -7,9 +7,16 @@
 * Metric can be calculated on cutouts ([#149](https://github.com/KIT-IAI/pyWATTS/issues/149))
 * Add custom scaler module for pre-processing ([#167](https://github.com/KIT-IAI/pyWATTS/issues/167))
 * Add a MASE Summary ([#148](https://github.com/KIT-IAI/pyWATTS/issues/148))
+* Improve Online Learning Functionality ([#117](https://github.com/KIT-IAI/pyWATTS/issues/117))
+  * Replace train_if callable by a ConditionObject. 
+  * Rename train_if to refit_condition.
+  * Separate transform from refit. I.e., the complete pipeline is transformed before any step is refitted.
+  * Add PeriodicCondition
+
 
 
 ### Changed
+* Retraining is triggered after all steps are transformed ([#117](https://github.com/KIT-IAI/pyWATTS/issues/117))
 
 ### Fixed
 * Try to handle non matching shapes in metrics. If not possible raise an exception.  ([#154](https://github.com/KIT-IAI/pyWATTS/issues/154))